+++ bits/01_version.js
@@ -1,1 +1,1 @@
-XLSX.version = '0.21.4';
+XLSX.version = '0.21.5';